+++ examples/public-dns-external/main.tf
@@ -25,10 +25,10 @@ module "wandb_infra" {
   database_snapshot_identifier = var.database_snapshot_identifier
   database_sort_buffer_size    = var.database_sort_buffer_size
 
-  allowed_inbound_cidr      = ["0.0.0.0/0"]
+  allowed_inbound_cidr      = var.allowed_inbound_cidr
   allowed_inbound_ipv6_cidr = ["::/0"]
 
-  eks_cluster_version            = "1.24"
+  eks_cluster_version            = "1.25"
   kubernetes_public_access       = true
   kubernetes_public_access_cidrs = ["0.0.0.0/0"]
 
@@ -51,7 +51,7 @@ data "aws_eks_cluster_auth" "app_cluster" {
 
 provider "kubernetes" {
   host                   = data.aws_eks_cluster.app_cluster.endpoint
-  cluster_ca_certificate = base64decode(data.aws_eks_cluster.app_cluster.certificate_authority.0.data)
+  cluster_ca_certificate = base64decode(data.aws_eks_cluster.app_cluster.certificate_authority[0].data)
   token                  = data.aws_eks_cluster_auth.app_cluster.token
 }
 
@@ -66,6 +66,7 @@ module "wandb_app" {
   bucket_queue               = "internal://"
   bucket_kms_key_arn         = module.wandb_infra.kms_key_arn
   database_connection_string = "mysql://${module.wandb_infra.database_connection_string}"
+  redis_connection_string    = "redis://${module.wandb_infra.elasticache_connection_string}?tls=true&ttlInSeconds=604800"
 
   wandb_image   = var.wandb_image
   wandb_version = var.wandb_version