+++ modules/app_eks/iam-role-policies.tf
@@ -1,0 +1,13 @@
+
+data "aws_iam_policy_document" "node_assume" {
+  statement {
+    sid     = "aa1"
+    actions = ["sts:AssumeRole"]
+    effect  = "Allow"
+    principals {
+      type        = "Service"
+      identifiers = ["ec2.amazonaws.com"]
+    }
+  }
+}
+