+++ modules/database/main.tf
@@ -16,6 +16,7 @@ locals {
   parameter_cluster_name = "${var.namespace}-aurora-${local.engine_version_tag}-cluster-parameter-group"
 }
 
+#https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/AuroraMySQL.Reference.ParameterGroups.html#AuroraMySQL.Reference.Parameters.Instance
 resource "aws_db_parameter_group" "default" {
   name        = local.parameter_group_name
   family      = local.parameter_family
@@ -54,11 +55,17 @@ resource "aws_db_parameter_group" "default" {
     value = var.sort_buffer_size
   }
 
+  parameter {
+    name  = "innodb_lru_scan_depth"
+    value = var.innodb_lru_scan_depth
+  }
+
   lifecycle {
     ignore_changes = [description]
   }
 }
 
+# https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/AuroraMySQL.Reference.ParameterGroups.html#AuroraMySQL.Reference.Parameters.Cluster
 resource "aws_rds_cluster_parameter_group" "default" {
   name        = local.parameter_cluster_name
   family      = local.parameter_family
@@ -76,6 +83,12 @@ resource "aws_rds_cluster_parameter_group" "default" {
     apply_method = "pending-reboot"
   }
 
+  parameter {
+    name         = "binlog_row_image"
+    value        = var.binlog_row_image
+    apply_method = "pending-reboot"
+  }
+
   lifecycle {
     ignore_changes = [description]
   }
@@ -83,49 +96,50 @@ resource "aws_rds_cluster_parameter_group" "default" {
 
 module "aurora" {
   source  = "terraform-aws-modules/rds-aurora/aws"
-  version = "6.2.0"
+  version = "7.7.1"
 
-  name           = var.namespace
-  engine         = "aurora-mysql"
-  engine_version = var.engine_version
-
-  allow_major_version_upgrade = true
-
-  instance_class = var.instance_class
-  instances      = { 1 = {} }
-
-  autoscaling_enabled = false
-
-  deletion_protection = var.deletion_protection
-
-  vpc_id                 = var.vpc_id
-  db_subnet_group_name   = var.db_subnet_group_name
-  create_db_subnet_group = var.create_db_subnet_group
-  subnets                = var.subnets
-
-  snapshot_identifier = var.snapshot_identifier
-
-  database_name = var.database_name
-
-  create_security_group = true
-  allowed_cidr_blocks   = var.allowed_cidr_blocks
-
-  iam_database_authentication_enabled = false
-  master_username                     = var.master_username
-  master_password                     = local.master_password
+  allow_major_version_upgrade         = true
+  allowed_cidr_blocks                 = var.allowed_cidr_blocks
+  apply_immediately                   = true
+  autoscaling_enabled                 = false
+  backup_retention_period             = var.backup_retention_period
+  create_db_subnet_group              = var.create_db_subnet_group
   create_random_password              = false
+  create_security_group               = true
+  database_name                       = var.database_name
+  db_cluster_parameter_group_name     = aws_rds_cluster_parameter_group.default.id
+  db_parameter_group_name             = aws_db_parameter_group.default.id
+  db_subnet_group_name                = var.db_subnet_group_name
+  deletion_protection                 = var.deletion_protection
+  enabled_cloudwatch_logs_exports     = ["audit", "error", "general", "slowquery"]
+  engine                              = "aurora-mysql"
+  engine_version                      = var.engine_version
+  iam_database_authentication_enabled = false
+  iam_role_force_detach_policies      = true
+  iam_role_name                       = "${var.namespace}-aurora-monitoring"
+  instance_class                      = var.instance_class
+  instances                           = { 1 = {} }
+  kms_key_id                          = var.kms_key_arn
+  master_password                     = local.master_password
+  master_username                     = var.master_username
+  monitoring_interval                 = 15
+  name                                = var.namespace
+  ////////////////////////////////////////////////////////////////////////////////////////
+  // !!! note on performance insights !!!
+  // AWS offers 7 days of performance insights free. keeping them after this period
+  // incurs a per-vcpu cost. so we can keep them for 7 days and they're free
+  ////////////////////////////////////////////////////////////////////////////////////////
+  performance_insights_enabled          = true
+  performance_insights_kms_key_id       = (var.performance_insights_kms_key_arn == "" || var.performance_insights_kms_key_arn == null) ? var.kms_key_arn : var.performance_insights_kms_key_arn
+  performance_insights_retention_period = 7
+  preferred_backup_window               = var.preferred_backup_window
+  preferred_maintenance_window          = var.preferred_maintenance_window
+  security_group_tags                   = { "Namespace" : "${var.namespace}" }
+  skip_final_snapshot                   = true
+  snapshot_identifier                   = var.snapshot_identifier
+  storage_encrypted                     = true
+  subnets                               = var.subnets
+  vpc_id                                = var.vpc_id
 
-  storage_encrypted = true
-  kms_key_id        = var.kms_key_arn
 
-  apply_immediately   = true
-  skip_final_snapshot = true
-
-  backup_retention_period      = var.backup_retention_period
-  preferred_backup_window      = var.preferred_backup_window
-  preferred_maintenance_window = var.preferred_maintenance_window
-
-  db_parameter_group_name         = aws_db_parameter_group.default.id
-  db_cluster_parameter_group_name = aws_rds_cluster_parameter_group.default.id
-  enabled_cloudwatch_logs_exports = ["audit", "error", "general", "slowquery"]
 }