+++ modules/database/variables.tf
@@ -3,6 +3,11 @@ variable "kms_key_arn" {
   type        = string
 }
 
+variable "performance_insights_kms_key_arn" {
+  description = "Specifies an existing KMS key ARN to encrypt the performance insights data if performance_insights_enabled is was enabled out of band"
+  type        = string
+}
+
 variable "namespace" {
   type        = string
   description = "The name prefix for all resources created."
@@ -102,3 +107,17 @@ variable "master_username" {
   default     = "wandb"
 }
 
+# DB Instance Parameters
+variable "innodb_lru_scan_depth" {
+  description = "Specifies the innodb_lru_scan_depth value to set for the database"
+  type        = number
+  default     = 128
+}
+
+
+# Cluster parametes
+variable "binlog_row_image" {
+  description = "Value for binlog_row_image"
+  type        = string
+  default     = "minimal"
+}