+++ modules/redis/main.tf
@@ -3,12 +3,12 @@ locals {
 }
 
 resource "aws_elasticache_replication_group" "default" {
-  replication_group_id          = "${var.namespace}-rep-group"
-  replication_group_description = "${var.namespace}-rep-group"
-  number_cache_clusters         = 2
-  port                          = 6379
+  replication_group_id = "${var.namespace}-rep-group"
+  description          = "${var.namespace}-rep-group"
+  num_cache_clusters   = 2
+  port                 = 6379
 
-  node_type            = "cache.t2.medium"
+  node_type            = var.node_type
   parameter_group_name = "default.redis6.x"
   engine_version       = local.redis_version